T20 - Virat
Data is taken from the commentary at http://cricsheet.org/ for every ball across all the T-20 matches in IPL. The information is then parsed in python and store into a pandas data-frame for the purpose of cleansing the data and extracting useful features from it. All the cleaning and pre-processing is done using the pandas data frame. Model is generated and tested from the historical data of Batsman Virat Kohli, and the same model is used to predict the score of Rohit Sharma. 10 features and 1 label is generated from this step and the data is exported into a CSV format to be used in Weka for data Analysis.